+++ pkgs/ffigen/CHANGELOG.md
@@ -4,6 +4,7 @@
   is linked with `-dead_strip`.
 - Handle dart typedefs in import/export of symbol files.
 - Add support for blocking ObjC blocks that can be invoked from any thread.
+- Add support for blocking ObjC protocol methods.
 
 ## 16.0.0